pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "data-cbs-interview"
version = "0.1.0"
description = "DataOps pipeline for financial transaction data"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.34",
    "requests>=2.31",
    "yfinance>=0.2",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "pyarrow>=15.0",
    "duckdb==1.4.4",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py311"
